# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-First CRM HCP Module API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type HcpTier = (typeof HcpTier)[keyof typeof HcpTier];

export const HcpTier = {
  A: "A",
  B: "B",
  C: "C",
} as const;

export interface Hcp {
  id: number;
  firstName: string;
  lastName: string;
  specialty: string;
  institution: string;
  territory: string;
  email?: string | null;
  phone?: string | null;
  npi?: string | null;
  tier: HcpTier;
  lastInteractionDate?: string | null;
  totalInteractions: number;
  createdAt: string;
  updatedAt: string;
}

export type HcpCreateInputTier =
  (typeof HcpCreateInputTier)[keyof typeof HcpCreateInputTier];

export const HcpCreateInputTier = {
  A: "A",
  B: "B",
  C: "C",
} as const;

export interface HcpCreateInput {
  firstName: string;
  lastName: string;
  specialty: string;
  institution: string;
  territory: string;
  email?: string | null;
  phone?: string | null;
  npi?: string | null;
  tier?: HcpCreateInputTier;
}

export interface HcpListResponse {
  hcps: Hcp[];
  total: number;
  limit: number;
  offset: number;
}

export type InteractionType =
  (typeof InteractionType)[keyof typeof InteractionType];

export const InteractionType = {
  visit: "visit",
  call: "call",
  email: "email",
  virtual_meeting: "virtual_meeting",
  conference: "conference",
  other: "other",
} as const;

export type InteractionSentiment =
  | (typeof InteractionSentiment)[keyof typeof InteractionSentiment]
  | null;

export const InteractionSentiment = {
  positive: "positive",
  neutral: "neutral",
  negative: "negative",
} as const;

export interface Interaction {
  id: number;
  hcpId: number;
  hcpName: string;
  type: InteractionType;
  date: string;
  /** Duration in minutes */
  duration?: number | null;
  notes: string;
  aiSummary?: string | null;
  sentiment?: InteractionSentiment;
  productsDiscussed: string[];
  followUpRequired: boolean;
  followUpDate?: string | null;
  followUpNotes?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type InteractionCreateInputType =
  (typeof InteractionCreateInputType)[keyof typeof InteractionCreateInputType];

export const InteractionCreateInputType = {
  visit: "visit",
  call: "call",
  email: "email",
  virtual_meeting: "virtual_meeting",
  conference: "conference",
  other: "other",
} as const;

export interface InteractionCreateInput {
  hcpId: number;
  type: InteractionCreateInputType;
  date: string;
  duration?: number | null;
  notes: string;
  productsDiscussed?: string[];
  followUpRequired?: boolean;
  followUpDate?: string | null;
  followUpNotes?: string | null;
}

export type InteractionUpdateInputType =
  (typeof InteractionUpdateInputType)[keyof typeof InteractionUpdateInputType];

export const InteractionUpdateInputType = {
  visit: "visit",
  call: "call",
  email: "email",
  virtual_meeting: "virtual_meeting",
  conference: "conference",
  other: "other",
} as const;

export interface InteractionUpdateInput {
  type?: InteractionUpdateInputType;
  date?: string;
  duration?: number | null;
  notes?: string;
  productsDiscussed?: string[];
  followUpRequired?: boolean;
  followUpDate?: string | null;
  followUpNotes?: string | null;
}

export interface InteractionListResponse {
  interactions: Interaction[];
  total: number;
  limit: number;
  offset: number;
}

export type DashboardSummaryByTypeItem = {
  type: string;
  count: number;
};

export type DashboardSummaryBySpecialtyItem = {
  specialty: string;
  count: number;
};

export interface DashboardSummary {
  totalHcps: number;
  totalInteractions: number;
  interactionsThisMonth: number;
  interactionsThisWeek: number;
  pendingFollowUps: number;
  byType: DashboardSummaryByTypeItem[];
  bySpecialty: DashboardSummaryBySpecialtyItem[];
  recentInteractions: Interaction[];
}

export interface AgentRequest {
  message: string;
  sessionId?: string | null;
  hcpId?: number | null;
  interactionId?: number | null;
}

export type AgentResponseExtractedData = { [key: string]: unknown } | null;

export interface AgentResponse {
  response: string;
  toolsUsed: string[];
  extractedData?: AgentResponseExtractedData;
  sessionId: string;
}

export type ChatMessageRole =
  (typeof ChatMessageRole)[keyof typeof ChatMessageRole];

export const ChatMessageRole = {
  user: "user",
  assistant: "assistant",
} as const;

export interface ChatMessage {
  role: ChatMessageRole;
  content: string;
}

export interface ChatRequest {
  messages: ChatMessage[];
  hcpId?: number | null;
  sessionId?: string | null;
}

export interface ChatResponse {
  message: string;
  extractedInteraction?: InteractionCreateInput | null;
  sessionId: string;
  toolsUsed: string[];
  readyToLog: boolean;
}

export type ListHcpsParams = {
  /**
   * Search by name, specialty, or institution
   */
  search?: string;
  /**
   * Filter by specialty
   */
  specialty?: string;
  /**
   * Filter by territory
   */
  territory?: string;
  limit?: number;
  offset?: number;
};

export type ListInteractionsParams = {
  /**
   * Filter by HCP ID
   */
  hcpId?: number;
  /**
   * Filter by interaction type
   */
  type?: string;
  limit?: number;
  offset?: number;
};