# lib/api-zod/src/generated/types/agentRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-First CRM HCP Module API
 * OpenAPI spec version: 0.1.0
 */

export interface AgentRequest {
  message: string;
  sessionId?: string | null;
  hcpId?: number | null;
  interactionId?: number | null;
}