# lib/api-zod/src/generated/types/agentResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-First CRM HCP Module API
 * OpenAPI spec version: 0.1.0
 */
import type { AgentResponseExtractedData } from "./agentResponseExtractedData";

export interface AgentResponse {
  response: string;
  toolsUsed: string[];
  extractedData?: AgentResponseExtractedData;
  sessionId: string;
}