# lib/api-zod/src/generated/types/chatResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-First CRM HCP Module API
 * OpenAPI spec version: 0.1.0
 */
import type { InteractionCreateInput } from "./interactionCreateInput";

export interface ChatResponse {
  message: string;
  extractedInteraction?: InteractionCreateInput | null;
  sessionId: string;
  toolsUsed: string[];
  readyToLog: boolean;
}