# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-First CRM HCP Module API
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryBySpecialtyItem } from "./dashboardSummaryBySpecialtyItem";
import type { DashboardSummaryByTypeItem } from "./dashboardSummaryByTypeItem";
import type { Interaction } from "./interaction";

export interface DashboardSummary {
  totalHcps: number;
  totalInteractions: number;
  interactionsThisMonth: number;
  interactionsThisWeek: number;
  pendingFollowUps: number;
  byType: DashboardSummaryByTypeItem[];
  bySpecialty: DashboardSummaryBySpecialtyItem[];
  recentInteractions: Interaction[];
}