# lib/api-zod/src/generated/types/hcp.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-First CRM HCP Module API
 * OpenAPI spec version: 0.1.0
 */
import type { HcpTier } from "./hcpTier";

export interface Hcp {
  id: number;
  firstName: string;
  lastName: string;
  specialty: string;
  institution: string;
  territory: string;
  email?: string | null;
  phone?: string | null;
  npi?: string | null;
  tier: HcpTier;
  lastInteractionDate?: Date | null;
  totalInteractions: number;
  createdAt: Date;
  updatedAt: Date;
}