# lib/api-zod/src/generated/types/hcpListResponse.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-First CRM HCP Module API
 * OpenAPI spec version: 0.1.0
 */
import type { Hcp } from "./hcp";

export interface HcpListResponse {
  hcps: Hcp[];
  total: number;
  limit: number;
  offset: number;
}