# lib/api-zod/src/generated/types/interactionCreateInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-First CRM HCP Module API
 * OpenAPI spec version: 0.1.0
 */
import type { InteractionCreateInputType } from "./interactionCreateInputType";

export interface InteractionCreateInput {
  hcpId: number;
  type: InteractionCreateInputType;
  date: Date;
  duration?: number | null;
  notes: string;
  productsDiscussed?: string[];
  followUpRequired?: boolean;
  followUpDate?: Date | null;
  followUpNotes?: string | null;
}