# lib/api-zod/src/generated/types/interactionUpdateInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * AI-First CRM HCP Module API
 * OpenAPI spec version: 0.1.0
 */
import type { InteractionUpdateInputType } from "./interactionUpdateInputType";

export interface InteractionUpdateInput {
  type?: InteractionUpdateInputType;
  date?: Date;
  duration?: number | null;
  notes?: string;
  productsDiscussed?: string[];
  followUpRequired?: boolean;
  followUpDate?: Date | null;
  followUpNotes?: string | null;
}